Revert "refactor[api-server]: 파츠 패키지 상위 패키지로 위치 변경"
This reverts commit a00459d0c671191a9f20d5b058c21df15052a957.
