+++ src/main/java/com/watermelon/server/randomevent/parts/domain/LotteryApplierParts.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.parts.domain;
+package com.watermelon.server.randomevent.parts.domain;
 
 import com.watermelon.server.BaseEntity;
 import com.watermelon.server.randomevent.domain.LotteryApplier;
