+++ src/main/java/com/watermelon/server/randomevent/parts/domain/Parts.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.parts.domain;
+package com.watermelon.server.randomevent.parts.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
