+++ src/main/java/com/watermelon/server/randomevent/parts/exception/PartsDrawLimitExceededException.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.parts.exception;
+package com.watermelon.server.randomevent.parts.exception;
 
 public class PartsDrawLimitExceededException extends RuntimeException {
 }
