+++ src/main/java/com/watermelon/server/randomevent/parts/repository/PartsRepository.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.parts.repository;
+package com.watermelon.server.randomevent.parts.repository;
 
-import com.watermelon.server.parts.domain.Parts;
+import com.watermelon.server.randomevent.parts.domain.Parts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PartsRepository extends JpaRepository<Parts, Long> {
