+++ src/main/java/com/watermelon/server/randomevent/parts/repository/PartsRewardRepository.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.parts.repository;
+package com.watermelon.server.randomevent.parts.repository;
 
-import com.watermelon.server.parts.domain.PartsReward;
+import com.watermelon.server.randomevent.parts.domain.PartsReward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PartsRewardRepository extends JpaRepository<PartsReward, Long> {
