+++ src/main/java/com/watermelon/server/randomevent/parts/service/LotteryApplierPartsService.java
@@ -1,8 +1,8 @@
-package com.watermelon.server.parts.service;
+package com.watermelon.server.randomevent.parts.service;
 
-import com.watermelon.server.parts.domain.Parts;
 import com.watermelon.server.randomevent.domain.LotteryApplier;
-import com.watermelon.server.parts.domain.LotteryApplierParts;
+import com.watermelon.server.randomevent.parts.domain.LotteryApplierParts;
+import com.watermelon.server.randomevent.parts.domain.Parts;
 
 import java.util.List;
