+++ src/main/java/com/watermelon/server/randomevent/parts/service/LotteryApplierPartsServiceImpl.java
@@ -1,9 +1,9 @@
-package com.watermelon.server.parts.service;
+package com.watermelon.server.randomevent.parts.service;
 
 import com.watermelon.server.randomevent.domain.LotteryApplier;
-import com.watermelon.server.parts.domain.LotteryApplierParts;
-import com.watermelon.server.parts.domain.Parts;
-import com.watermelon.server.parts.repository.LotteryApplierPartsRepository;
+import com.watermelon.server.randomevent.parts.domain.LotteryApplierParts;
+import com.watermelon.server.randomevent.parts.domain.Parts;
+import com.watermelon.server.randomevent.parts.repository.LotteryApplierPartsRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
