+++ src/main/java/com/watermelon/server/randomevent/parts/service/PartsServiceImpl.java
@@ -1,14 +1,14 @@
-package com.watermelon.server.parts.service;
+package com.watermelon.server.randomevent.parts.service;
 
-import com.watermelon.server.parts.domain.Parts;
-import com.watermelon.server.parts.dto.response.ResponseMyPartsListDto;
-import com.watermelon.server.parts.dto.response.ResponsePartsDrawDto;
-import com.watermelon.server.parts.dto.response.ResponseRemainChanceDto;
-import com.watermelon.server.parts.exception.PartsNotExistException;
-import com.watermelon.server.parts.repository.PartsRepository;
 import com.watermelon.server.randomevent.domain.LotteryApplier;
 import com.watermelon.server.randomevent.link.service.LinkService;
-import com.watermelon.server.parts.domain.LotteryApplierParts;
+import com.watermelon.server.randomevent.parts.domain.LotteryApplierParts;
+import com.watermelon.server.randomevent.parts.domain.Parts;
+import com.watermelon.server.randomevent.parts.dto.response.ResponseMyPartsListDto;
+import com.watermelon.server.randomevent.parts.dto.response.ResponsePartsDrawDto;
+import com.watermelon.server.randomevent.parts.dto.response.ResponseRemainChanceDto;
+import com.watermelon.server.randomevent.parts.exception.PartsNotExistException;
+import com.watermelon.server.randomevent.parts.repository.PartsRepository;
 import com.watermelon.server.randomevent.service.LotteryService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
