+++ src/main/java/com/watermelon/server/randomevent/service/LotteryServiceImpl.java
@@ -10,8 +10,8 @@ import com.watermelon.server.randomevent.dto.response.ResponseLotteryRankDto;
 import com.watermelon.server.randomevent.dto.response.ResponseLotteryWinnerDto;
 import com.watermelon.server.randomevent.dto.response.ResponseLotteryWinnerInfoDto;
 import com.watermelon.server.randomevent.dto.response.ResponseRewardInfoDto;
-import com.watermelon.server.parts.domain.PartsReward;
-import com.watermelon.server.parts.repository.PartsRewardRepository;
+import com.watermelon.server.randomevent.parts.domain.PartsReward;
+import com.watermelon.server.randomevent.parts.repository.PartsRewardRepository;
 import com.watermelon.server.randomevent.repository.LotteryApplierRepository;
 import com.watermelon.server.randomevent.repository.LotteryRewardRepository;
 import jakarta.transaction.Transactional;
